+++ client/landing/stepper/declarative-flow/internals/global.scss
@@ -89,7 +89,7 @@ button {
 .setup-blog,
 .celebration-step,
 .use-my-domain,
-.bulk-domain-transfer,
+.domain-transfer,
 .update-design {
 	padding: 60px 0 0;
 	box-sizing: border-box;
