+++ client/landing/stepper/declarative-flow/internals/steps-repository/domain-transfer-intro/index.tsx
@@ -2,14 +2,13 @@ import { useI18n } from '@wordpress/react-i18n';
 import { StepContainer } from 'calypso/../packages/onboarding/src';
 import FormattedHeader from 'calypso/components/formatted-header';
 import { recordTracksEvent } from 'calypso/lib/analytics/tracks';
-import CalypsoShoppingCartProvider from 'calypso/my-sites/checkout/calypso-shopping-cart-provider';
-import TransferDomains from './domains';
+import IntroStep from './intro';
 import type { Step } from '../../types';
 
 import './styles.scss';
 
-const Intro: Step = function Intro( { navigation, flow } ) {
-	const { submit, goBack } = navigation;
+const Intro: Step = function Intro( { navigation } ) {
+	const { submit } = navigation;
 	const { __ } = useI18n();
 
 	const handleSubmit = () => {
@@ -17,23 +16,17 @@ const Intro: Step = function Intro( { navigation, flow } ) {
 	};
 	return (
 		<StepContainer
-			flowName={ flow }
-			stepName="domains"
-			goBack={ goBack }
-			isHorizontalLayout={ false }
+			hideBack
+			stepName="intro"
 			isLargeSkipLayout={ false }
 			formattedHeader={
 				<FormattedHeader
-					id="bulk-domains-header"
-					headerText={ __( 'Add unlocked domains' ) }
-					align="left"
+					id="domain-transfer-header"
+					headerText={ __( 'Transfer domains' ) }
+					subHeaderText={ __( 'Follow along these 3 steps to transfer domains to WordPress.com' ) }
 				/>
 			}
-			stepContent={
-				<CalypsoShoppingCartProvider>
-					<TransferDomains onSubmit={ handleSubmit } />
-				</CalypsoShoppingCartProvider>
-			}
+			stepContent={ <IntroStep onSubmit={ handleSubmit } /> }
 			recordTracksEvent={ recordTracksEvent }
 			showHeaderJetpackPowered={ false }
 			showHeaderWooCommercePowered={ false }
